Rename: src/index.tsxをsrc/ScreenSwitcher.tsxに変更
